Move all services related to data management to data package
